+++ patzilla-ui/lib/mdc/snackbar.js
@@ -1,0 +1,56 @@
+// -*- coding: utf-8 -*-
+// (c) 2018 Andreas Motl <andreas.motl@ip-tools.org>
+'use strict';
+
+import { classes } from 'patzilla.lib.es6';
+import { MDCSnackbar } from '@material/snackbar';
+import { DirectRenderMixin } from 'patzilla.lib.marionette';
+
+export { SnackbarWidget };
+
+
+// Acquire CSS
+require('@material/snackbar/mdc-snackbar');
+
+class SnackbarWidget extends classes.many(Backbone.Marionette.ItemView, DirectRenderMixin) {
+    /*
+     * SnackbarWidget encapsulates the Snackbox Material Design Component (MDC)
+     * into a `Marionette.ItemView`.
+     *
+     * https://material-components.github.io/material-components-web-catalog/#/component/snackbar
+     *
+     * Details
+     * =======
+     *
+     * As this is intended to be instantiated only once, it will actually
+     * render just once and attach itself to the HTML body on the DOM element
+     * level.
+     *
+     * As the element is declared `aria-hidden="true"`, it will _not_ be
+     * displayed on the first hand.
+     *
+     * When used, it will show up and disappear at runtime, which is completely
+     * the responsibility of the `MDCSnackbar` controller component.
+     *
+    **/
+
+    // The HTML template for a MDC Snackbar
+    get template() {
+        return require('./snackbar.html');
+    }
+
+    initialize() {
+        this.render();
+    }
+
+    onRender() {
+        this.snackbar = new MDCSnackbar(this.el);
+        $("body").append(this.el);
+    }
+
+    show(message, options) {
+        //this.snackbar.show({message: message, timeout: 40000});
+        this.snackbar.show({message: message});
+    }
+
+}