+++ CHANGES.rst
@@ -6,6 +6,7 @@ IP Navigator CHANGES
 Development
 ===========
 - [ui] Gradually ramping up ES6
+- [ui] Notify users by MDC Snackbar, ditch NotificationFx
 
 
 2018-11-26 0.161.1